+++ lib/circleci/coverage_reporter/artifact.rb
@@ -4,11 +4,13 @@ module CircleCI
     #
     # @attr path [String] abstract path to the artifact in CircleCI container
     # @attr url [String] URL of the artifact
-    Artifact = Struct.new(:path, :url) do
+    # @attr node_index [Integer] the ID of the artifact's container
+    Artifact = Struct.new(:path, :url, :node_index) do
       # @param value [String]
+      # @param node_index [Integer, nil]
       # @return [Boolean]
-      def match?(value)
-        path.end_with?(value)
+      def match?(value, node_index: nil)
+        path.end_with?(value) && (node_index.nil? || self.node_index == node_index)
       end
 
       # @return [String] content of the artifact